Sadmin - Admin Service - Update create new admin function

+++ src/services/adminServices.ts
@@ -120,12 +120,10 @@ export const deleteAdmin = async (req: Request, res: Response) => {
   const { adminId } = req.params;
 
   try {
-    const admin = await User.findByIdAndDelete(adminId);
-
+    const admin = await User.findByIdAndDelete({userID: adminId, role: "admin"});
     if (!admin) {
       return res.status(404).json({ message: "Admin not found" });
     }
-
     res.status(201).json({ message: "Admin deleted successfully" });
   } catch (error) {
     res.status(400).json(error);
